internally prefix every M4 macro with format and visibility
Prefixing the macros is a prerequisite for allowing the user to
overwrite them in a single ~/.m4rkup.m4 file (or similar), to be
implemented in the future.

For the user, this change is transparent due to the fact that
lib/bootstrap.m4, which now does the main work, loads the according
output format libraries and defines the public interface macros as stubs
which call the prefixed versions of the respective output library.

./minutes-m4rkup suddenly got a lot simpler due to the fact that m4
looks for `bootstrap.m4` in its include paths, basically doing the work
of get_lib_files().